requirements.txt
selenium
langchain
ollama
python-dotenv
pydantic
asyncio
pyyaml
